+++ tests/rdl_src/compile_fail/access_modes_hardware/no_sw_reg.rs
@@ -1,0 +1,12 @@
+use access_modes_hardware::AccessModesTest;
+use peakrdl_rust::io::MockIO;
+
+const SIZE: usize = AccessModesTest::<()>::SIZE;
+
+fn main() {
+    let memory: MockIO<SIZE> = MockIO::new_zeroed();
+    let top = unsafe { AccessModesTest::from_ptr_with(memory.base_ptr(), &memory) };
+
+    // this register shouldn't even be generated since all fields are hw=na
+    let sw_reg = top.sw_reg();
+}