+++ tests/rdl_src/compile_fail/access_modes_read_only/no_write_method.rs
@@ -1,0 +1,12 @@
+use access_modes_read_only::AccessModesTest;
+use peakrdl_rust::io::MockIO;
+
+const SIZE: usize = AccessModesTest::<()>::SIZE;
+
+fn main() {
+    let memory: MockIO<SIZE> = MockIO::new_zeroed();
+    let top = unsafe { AccessModesTest::from_ptr_with(memory.base_ptr(), &memory) };
+
+    // the write method shouldn't be implemented since read_only=true
+    top.sw_reg().write(|reg| {});
+}